refactor: fix unset file size limit in settings, use openExternal folder
