+++ public/app/listeners/openPathInExplorer.js
@@ -6,7 +6,7 @@ const logger = require('../logger');
 const openPathInExplorer = () => (event, location) => {
   logger.debug(`showing path ${location} in explorer`);
   // path.resolve necessary for windows location path
-  shell.showItemInFolder(path.resolve(location)); // Show the given file in a file manager. If possible, select the file.
+  shell.openExternal(`file://${path.resolve(location)}`); // Show the given file in a file manager. If possible, select the file.
 };
 
 module.exports = openPathInExplorer;
