+++ src/config/constants.js
@@ -98,9 +98,6 @@ export const UTILS_FILES = {
 
 export const CIRCLE_PROGRESS_SIZE = 26;
 
-export const FILE_SIZE_LIMIT_OPTIONS = [0, 500, 1000, 2500, 5000, 10000];
-export const DEFAULT_FILE_SIZE_LIMIT = FILE_SIZE_LIMIT_OPTIONS[2];
-
 export const PARAMETER_TYPES_PYTHON = {
   [PARAMETER_TYPES.FIELD_SELECTOR]: 'str',
   [PARAMETER_TYPES.FLOAT_INPUT]: 'float',
