+++ src/shared/constants.js
@@ -41,6 +41,9 @@ const DATASET_TYPES = {
 
 const EXECUTION_NAME_MAX_LENGTH = 25;
 
+const FILE_SIZE_LIMIT_OPTIONS = [0, 500, 1000, 2500, 5000, 10000];
+const DEFAULT_FILE_SIZE_LIMIT = FILE_SIZE_LIMIT_OPTIONS[2];
+
 module.exports = {
   ALGORITHM_TYPES,
   APP_BACKGROUND_COLOR,
@@ -51,4 +54,6 @@ module.exports = {
   EXECUTION_STATUSES,
   DATASET_TYPES,
   EXECUTION_NAME_MAX_LENGTH,
+  FILE_SIZE_LIMIT_OPTIONS,
+  DEFAULT_FILE_SIZE_LIMIT,
 };
